Plug hole in sync rewards between finalized and archived epochs
When sync rewards are requested, we can either go to the antiquary if
we are running an archive caplin node, or use the forkchoiceStore to
answer the request.

Fix the selection logic between these two paths, so the choice is
taken based on the availability in forkchoiceStore, instead of
finality.  Since some recent optimizations, the finality doesn't
always match the forkchoiceStore availability anymore.