txpool: addRemoteTxs broadcast - missed some txs (#13543)
The txpool does not broadcast remote transactions or send them to their
peers. Steps to reproduce the bug:

1) The devnet contains 3 nodes: node1 is connected to node2 via P2P, and
node2 is connected to node3 via P2P.
2) Send a transaction to node3, 
3) It is added to node3 as a local transaction and broadcasted to node2.
4) Node2 receives it as a remote transaction, adds it to its pool, but
never broadcasts it to its peers.

Is this the expected behavior, or is it a bug?

If it is a bug, it seems we may have forgotten to promote transactions
added to the pool (similar to how local transactions are handled in the
AddLocalTxns function).