polygon/sync: retry ufc on busy response (#13640)
a user on discord flagged the following issue (discord
[link](https://discord.com/channels/687972960811745322/983710221308416010/1334851657355235358)):
```
[INFO] [01-31|10:44:40.015] [sync] update fork choice done           in=807.13624ms
[INFO] [01-31|10:44:40.022] [sync] update fork choice                block=67365147 hash=0xb697ef9de67d61f300fda27367617ec270d9bcd15e9a7c73facdc253b6fc0273 age=3s
[EROR] [01-31|10:44:40.022] failed to update fork choice             latestValidHash=0x0000000000000000000000000000000000000000000000000000000000000000 err="fork choice update failure: status=5, validationErr=''"
[EROR] [01-31|10:44:40.038] runPostForkchoiceInBackground            error="[1/6 OtterSync] db closed"
[EROR] [01-31|10:44:43.029] polygon sync crashed - stopping node     err="pos sync failed: fork choice update failure: status=5, validationErr=''"
```

Note: `status=5` means `ExecutionStatus_Busy`

This can happen because there is a possibility that
`runPostForkchoiceInBackground` acquires the execution engine semaphore
after a successful UFC (line which says `update fork choice done
in=807.13624ms`) to do pruning. But very shortly after that there is
another call to UFC (in this case 7ms after - line which says `update
fork choice block=67365147`) which then fails with a `Busy` response
since it can't acquire the semaphore (take by
`runPostForkchoiceInBackground` for pruning).

This PR fixes this situation by adding retries on `Busy` response when
calling UFC. Such a retry already existed when calling `InsertBlocks` so
this is something that just got missed.