execution: fix and simplify block consumer exec delays (#13641)
Now that we are using Astrid by default it means that all of our block
consumers for all chains (Ethereum/Gnosis/Polygon) use the same flow for
execution - UpdateForkChoice.

This means we can simplify when/where we call
`UpdateBlockConsumerPreExecutionDelay` and
`UpdateBlockConsumerPostExecutionDelay`.

This should also resolve the inconsistencies that @mriccobene has seen
between Polygon/Ethereum.