Fixed clique consensus mining process (#13723)
We encountered an issue with producing blocks in the devnet environment
when the --dev.period flag is set to 0, which is its default value. The
issue was caused by an infinite wait on a channel read in one of the
goroutines inside the `StartMining` function.

Additionally, in this PR, I removed the unused `noempty` variable from
the `stage_mining_exec.go` file. The logic remains unchanged—I only
removed if blocks that were never executed.