polygon: Add Cloudflare transient err (#13731)
Polygon node crashed with the following error

```
[EROR] [02-06|09:17:35.922] polygon sync crashed - stopping node     err="pos sync heimdall failed: span scraper failed: can't fetch id range: error while fetching data from Heimdall: url='https://heimdall-api-amoy.polygon.technology/bor/latest-span', status=404, body='<!DOCTYPE html>\n<html>\n\n<head>\n  <title>Error ・ Cloudflare Access</title>
```

This should not be a failing error as it is likely due to Cloudflare
masking downtime of the Heimdall endpoint that are currently covered by
`ErrServiceUnavailable`. We specifically check for `404` status codes as
other status codes like `403` may signal other errors such as
Cloudflare's DDoS protection.