shutter: port over crypto pkg (#13752)
part 2 of https://github.com/erigontech/erigon/issues/13384
used in https://github.com/erigontech/erigon/pull/13725

Short-term ported from:
https://github.com/shutter-network/shutter/tree/main/shlib/shcrypto
Long term plan is to import this as a library from the shutter repos,
however for now im doing this port because currently that library has
go-ethereum dependencies and so that will bring go-ethereum transitive
dependencies to erigon which is not desirable. I've had a look and
actually it is possible to remove that go-ethereum dependency in the
shutter crypto package but needs a bit of work and cooperation with the
shutter team. I've created an issue
(https://github.com/erigontech/erigon/issues/13724) for this and have
left it for a future time (after we have a working shutter validator
implementation).