silkworm: add state snapshots (#13766)
Previously the Silkworm library was capable of reading Erigon 2 "blocks"
snapshots (.seg and .idx). The code to provide them to Silkworm was a
part of RoSnapshots. The old code is moved to be a part of
silkworm.SnapshotsRepository, and migrated to the updated Silkworm C
API.

The new code adds an ability to provide Erigon 3 "state" snapshots for
reading by Silkworm during evmone execution mode (`--silkworm.exec`). It
is based on the AggregatorRoTx.StaticFilesInRange method that was made
public (with the related DTOs) to be accessible from the silkworm
package.

BtIndex.DataHandle() added to expose the memory mapped file pointer to
Silkworm (in the same way as Decompressor and recsplit.Index). A bunch
of public methods added to filesItem via FilesItem interface to expose
the snapshot file objects to Silkworm.