types: Move wrapped blob txn marshal to different func (#14084)
Restores `MarshalBinary` contents from before
https://github.com/erigontech/erigon/pull/13975
The full marshal of wrapped blobs are only used in a test, hence moved
the logic to `MarshalBinaryWrapped`. This does make this different from
other transaction types, as it only returns the marshalled bytes of the
internal

Fixes an issue in which wrapped blobs were getting included to the
returned payload.