improvement(log): add call stack to func on the record (#14052)
### Description

Moved the stack.Caller to a func on the Record. Calling the stack
multiple times every log gets expensive even when you are logging under
the level defined in config. This is because the call stack will get
created every time the Log() func runs on the handlers, even if the call
stack is not being used.

### Context

This issue came up in CDK Erigon saying the log was performing poorly in
a pprof benchmark.

Issue 1:
[1637](https://github.com/0xPolygonHermez/cdk-erigon/issues/1637)
Issue 2:
[1770](https://github.com/0xPolygonHermez/cdk-erigon/issues/1770)

After running a bechmark pprof after this changeset, the call stack has
gone from the flame graph and improves performance.

### CDK Erigon changeset


[1f4cca638da71ffec4490bf21fa84004984b3546](https://github.com/gateway-fm/log/commit/1f4cca638da71ffec4490bf21fa84004984b3546)