Downgrade gopsutils version (#14477)
Downgrade of the gopsutils dependency due to a regression introduced in
version v4.24.9.

After upgrading to gopsutils v4.24.9, we encountered issues detailed in
[Erigon Issue
#14415](https://github.com/erigontech/erigon/issues/14415). The problem
appears to be introduced specifically in that release.

To maintain stability and avoid the regression, we are reverting
gopsutils to the last known working version until a fix is introduced
upstream.