Invalidate potentially stale header hash (#14485)
There are scenarios where we copy an immutable header, apply changes to
it, then calculate the new hash, but because the cached value is also
copied over in `CopyHeader()`, then we end up calculating a stale hash.

This could happen due to the modifications in `NewBlock()` which
invalidate the copied over cached hash.

Likewise `*Block.WithSeal()` makes a mutable header immutable, however
it doesn't ensure to refresh the cached hash.

These scenarios are covered in this PR by refreshing the cached header
hash.

---------

Co-authored-by: antonis19 <antonis19@users.noreply.github.com>