Provide 'totalGasLimit' to the 'ValidateAATransaction' gas pool; comment out sub gas (#14407)
Notice that succesfull pre-charging the entire transaction gas limit is
part of the transaction validity condition.
Also, as this empties the gas pool, it is not clear how to re-charge the
gas for individual frames in that scenario. I assume only one of the gas
subtractions can be kept, while the other should be discarded.