tests: use `go test -short` instead of build tag `integration` (#14662)
Use the [standard
approach](https://stackoverflow.com/questions/55180613/how-do-i-write-test-in-go-that-make-use-of-the-short-flag-and-can-it-be-combine)
instead of `go:build integration`.

To skip a slow test y default, add
```
if testing.Short() {
  t.Skip()
}
```
at the beginning of a test.

Use `make test` to run short tests and `make test-all` to run all test.

See also Issue #10184.

On my MacBook the run time of `make test` is reduced to 1m09s from
2m:47s.