Switches CDN to R2 (#14829)
This is a companion PR to
https://github.com/erigontech/erigon-snapshot/pull/544

From multi CDN support, it switches it to R2.

Files are not published yet (for main and release/3.0) at the time I'm
opening this PR, so to test it you should add:
`SNAPS_GIT_BRANCH="michele-mirror-test"` env variable.

This is a minimalistic PR in order to quickly workaround
https://github.com/erigontech/erigon/issues/14808 ; other long term
improvements are required.