Fix/docker GO version (#15159)
**Fix Docker build by bumping Go version in builder image to
1.24.1-alpine3.20**

**Description:**
The Docker build was failing because the Go version in the builder image
was outdated compared to the Go version specified in `go.mod`.

In PR [#15112](https://github.com/erigontech/erigon/pull/15112), the Go
version in `go.mod` was bumped from 1.22 to 1.23, but the Dockerfile
still used `golang:1.22-alpine3.19` as the builder image. This mismatch
caused the build to break.

This PR updates the Dockerfile to use `golang:1.24.1-alpine3.20` as the
builder image, aligning it with the `main` branch Go version and
ensuring a successful build.