Add apply/async operations for transactions (#15186)
This change adds an Apply(tx) function to transactions and an AsyncTx
which uses a channel to pass the apply function between go-routines.

It is used to enable parallel workers to access the headers/blocks on
the main RW execution thread, even with an external tx whilst
maintaining their own read only transaction for state access.

There is an overhead in doing this - but so long as the main execution
thread remains live in its wait loop this does not impact processing due
to the relative rareness of these types of call from the evm.