ci: fix and enable TestEthSubscribe (#15198)
Fixes #14546 

The main task executing `TestEthSubscribe` sporadically gets blocked
waiting for new headers in case the anonymous task created and started
in `rpchelper.New` running the subscription loop has not yet reached the
`Subscribe` call, when `stages.StageLoopIteration` completes its
execution. In such a case, the internal subscription channels in
`Events` have not been created yet and thus the NEW_HEADER notifications
are not sent.

A simple solution, not touching the production code, is relying on the
initial NEW_SNAPSHOT notification that is sent unconditionally by
`EthBackendServer` after creating the internal subscription channels.