Mutli goroutine Shared Domains with external tx's  (#15240)
These are a set of changes/fixed to shared domains to make them operate
in a parallel processing environment where:

1.  Each worker holds it own RO mdbx transactions
2. Workers run in parallel - i.e. no Workerpool level logging
3. Storage is shared between workers and with a main goroutine which
performs/orchestrates commitment calculations.

To make the work it is not feasible to hold a single mdbx in the shared
domain and use SetTx to update it. This leads to the parallel works
tripping over each other or holding unnecessary lock across the domain.
Given that state reads are a critical resource for parallel processing
access in a `free threading` environment seems to maximize throughput.

After this merge transactions are held in Readers and Writers rather
than the shared domain. Not that a SD writer only needs a RoTx as its
updates the local maps in the domain prior to flushing - so in practice
so long as the map access is protected there is no need to a single
threaded RW transaction except for the Flush - which happens in a single
main goroutine.

The additional changes in this PR come from fixing the various thread
safety issues which arise from running in the mode described above as at
various stages of the previous versions of the code assumptions where
made about a single access to the overall structure - which is no longer
the case.

---------

Co-authored-by: alex.sharov <AskAlexSharov@gmail.com>