Add Tx Dependencies (#15415)
Bor passes TX dependencies in the ExtraData fields of its headers. These
values are populated by the block producer and can be used by all
clients to manage dependencies when doing parallel block execution.

Note that at the moment this dependency information is ignored by the
parallel execution because it appears to slow down execution. This needs
further analysis.