cp: cancel previous job runs in integration and qa-rpc (#15566)
maybe helps with recent tests scheduling flakiness

---
when new commits are pushed into PR, the job runs of previous commits of
the PR (in particular, qa-rpc and integration tests) should be
cancelled.
This already happens with ci test. Just adding it for other two
workflows.

It's just something to save unnecessary job runs.
NOTE: this doesn't effect job runs on main branch or release branches
(these are the branches on which the requirement to know "which commit
broke a test" is needed)