Fix datadir execution permissions for multi-user setups (#15594)
Per https://github.com/erigontech/devops-metarepo-for-issues/issues/80,
Erigon creates data directories without executable permissions
preventing group members from reading the contents.

https://github.com/erigontech/devops-metarepo-for-issues/issues/80 fixes
retroactively, but this PR will fix going forward too.

Note that umask 0o022 is the default on most systems and will mask newly
created directories to 0o755 which I believe is the intended default
(consider the trailing `4` in the original 0o764).

Further note that to have the behaviour intended by 0o764, the umask
would need to be modified to 0o002 but that's out of the scope of this
fix.

This should be backported to release as well.

(cherry picked from commit 4122f833cab250cc77fb214c9f7887036aa8aff3)