Don't add preverified hashes after initial sync completed (#15674)
This is to complete an item from
https://github.com/erigontech/erigon/pull/15043:

> Don't redownload files after preverified.toml is committed if they go
missing. This occurs due to merges etc.

At node startup, the preverified.toml that was committed locally from a
previous initial sync is scanned and the torrents in it are requested
from the downloader. If in a previous run, torrents were removed due to
merges, the torrents in the final snapshot will be added again (to my
understanding). I can see two ways to deal with this:

1. Mark manually removed torrents with a file like name.removed, or
2. Once the preverified.toml is committed, don't request any torrents
from it any longer.

The second option removes a lot of the logic from the downloader, where
it doesn't belong. This PR implements that. The only downside I can see
here is it makes it harder to manually repair a snapshot dir, for
example you could "resync" a dir by removing all the `.removed` files in
1. However you could also sync to the latest initial snapshot by
removing whatever preverified.toml you have, possibly wiping any
.torrent files, and let the node resync (it will reuse existing data now
too since #15043).

I'm working on a sequence diagram that should describe how it fits
together so it's clear and easy to work with the downloader snapshot dir
and to check this logic.