fix concurrency groups (CG) for workflow jobs (#15723)
this PR fixes concurrency groups (CG) for jobs.

current situation:
- in a bunch of places, something like `ci-${{ matrix.os }}-${{
github.ref }}` is used as CG;
- `cancel-in-progress` = false if release or main branch; else true

this became a problem with cancelling jobs on `main`, even though
`cancel-in-progress` was false for it. This was because:
1. the job on main was defined by CG example - `ci-${{ matrix.os }}-${{
github.ref }}`.
2. Successive commits on main have the same CG.
3. within a CG there can be at most one pending job + at most 1 running
job.

point 3. was not considered when the CG structure was decided, leading
to cancelling jobs on main. **I already did the change for
`test-all-erigon` workflow** -
https://github.com/erigontech/erigon/pull/15700. This stopped the
cancelling jobs on main for this particular workflow.

This PR fixes the CG for all jobs using CG.

---

essentially, now I changed the group to:

```
 ${{
          (github.ref == 'refs/heads/main' || startsWith(github.ref, 'refs/heads/release/')) &&
          format('{0}-{1}', github.workflow, github.run_id) ||
          format('{0}-{1}', github.workflow, github.ref)
        }}
```

- so for main and release, the jobs have the job's `run_id`, which makes
the CG unique and so cannot be cancelled by another job.
- for PRs, the jobs are keyed by workflow_id and PR name -- which means
successive commits on PR will cause cancellation.

issue: https://github.com/erigontech/erigon/issues/15670