Pull latest webseed performance improvements WIP (#15765)
Update to the latest anacrolix/torrent commit for webseed performance.

This moves performance from ~5-10MB/s up past the old limit of
~100-125MB/s to ~350+MB/s in my testing.

There's a few corner cases I'm working through but it seems solid I want
to unstick performance for CI and dev.

The changes in erigon are mostly just observations and updates to code
while testing. The majority is upstream in
https://github.com/anacrolix/torrent/compare/anacrolix...global-webseed-requests?expand=1.
Feel free to take a look.

- In particular it's slow to finish up small files as it does requests
on a timer.
- I haven't tested how it reacts to bad snapshots yet.
- I haven't prune the boundaries of request extents to avoid a small
amount of waste.
- Webseed requests don't terminate early to avoid unnecessary downloads,
they should just error out and then pick something else.

This PR also doubles as a check that CI passes.