qa_tests: one-liner for RPC Integration Tests (#15978)
Closes #15891

The main change is moving `rpc-tests` repository checkout to
`run_rpc_tests.sh`. Now such script can be used in isolation on any
machines (e.g. snapshotters) to run the RPC tests using exactly the same
`rpc-tests` version as in CI by doing:
```
.github/workflows/scripts/run_rpc_tests.sh
```

This is for Ethereum mainnet, when merged I'll do the same for Gnosis
and Polygon.