rpcdaemon: eth_getLogs(): add check on pruning (#16042)
If an eth_getLogs() request specifies neither an address nor topics, and
the requested range falls on a pruned state, an error is returned to the
client. While if the request includes an address and/or topics, no error
is returned; instead, an empty log array is provided, which means the
client is not directly informed of the problematic request context
(i.e., querying a pruned state).
This happens because reverse index returns en empty list therefore
becomes empty logs