Fix double registration if debugMux is set to default mux
We don't want the debug mux to fallback to the default HTTP mux because it only has one use for now.

Add a guard in case we do actually do this down the track though. And because there's no point registering the same thing twice if they're the same mux anyway.

We also reinstate the missing handler for the regular node invocation.