Fix bor_getAuthor by hash (#16121)
The header needs to be read by hash using `FullBlockReader` which has
access to snapshots and MDBX, instead of `rawdb` which has access only
to MDBX data.

This was fixed before when the input is a block number in
https://github.com/erigontech/erigon/pull/15622 and in
https://github.com/erigontech/erigon/pull/15630 .

This PR handles the case where the input is a block hash.

Issues fixed: 
- https://github.com/erigontech/erigon/issues/15285
- https://github.com/erigontech/erigon/issues/16020

Co-authored-by: antonis19 <antonis19@users.noreply.github.com>